feat(seo): complete phase 1 refactor - implement SSR wrappers, unique metadata for all pages, optimized sitemap, and keyword-rich alt texts for all banners
